Add files via upload
The goal of the project was to create a system that could identify the seven fundamental human emotions—neutral, happy, sad, surprise, fear, disgust, and anger—in natural settings. The project used Keras with Tensor Flow for model creation and training, OpenCV for real-time image captured and image processing, Python as the main language, and other tools as needed.
The model was trained using the Expression in-the-Wild (ExpW) dataset, provided by madam during class, which is a sizable collection of facial photos with varied expressions acquired from the internet. Only 55% from the total datasets were used. The photos were RGB, resized to a standard size (64x64 pixels), applying different preprocessing methods and the labels were transformed into one-hot vectors for the Convolutional Neural Network (CNN).